% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tex_value.R
\name{handle_file_name}
\alias{handle_file_name}
\title{Title}
\usage{
handle_file_name(file_name, path = NULL)
}
\arguments{
\item{file_name}{File name to create on disk. \code{file_name} should have
either a \code{.tex} extension or no extension at all.}

\item{path}{Path to the directory to save the file to. \code{path} and
\code{file_name} are combined to create fully qualified file name. If
\code{Null} and \code{file_name} does not specify a full file name defaults
to working directory.}
}
\value{
a complete file name (with path and extension)
}
\description{
Title
}
\keyword{internal}
