% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_lyx.R
\name{print_lyx}
\alias{print_lyx}
\title{Title}
\usage{
print_lyx(latexFile, path = NULL, currentName = NULL, nrows = NULL)
}
\arguments{
\item{latexFile}{a}

\item{path}{a}

\item{currentName}{ab}

\item{nrows}{ab}
}
\value{

}
\description{
Title
}
\examples{
a
}
