% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tex_value.R
\name{print_tex_value}
\alias{print_tex_value}
\title{Title}
\usage{
print_tex_value(file_name, path = NULL, names = NULL)
}
\arguments{
\item{file_name}{File name to create on disk. \code{file_name} should have
either a \code{.tex} extension or no extension at all.}

\item{path}{Path to the directory to save the file to. \code{path} and
\code{file_name} are combined to create fully qualified file name. If
\code{Null} and \code{file_name} does not specify a full file name defaults
to working directory.}

\item{names}{ab}
}
\value{

}
\description{
Title
}
\examples{
a
}
