% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tex_values.R
\name{remove_tex_value}
\alias{remove_tex_value}
\title{Title}
\usage{
remove_tex_value(names, file_name, path = NULL)
}
\arguments{
\item{names}{a}

\item{file_name}{a}

\item{path}{a}
}
\description{
Title
}
\examples{
a

}
