% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tex_value.R
\name{save_tex_value}
\alias{save_tex_value}
\title{Title}
\usage{
save_tex_value(
  values,
  names,
  file_name,
  path = NULL,
  translate = TRUE,
  digits = 2,
  percent = FALSE,
  accuracy = 1,
  override = TRUE
)
}
\arguments{
\item{values}{a}

\item{names}{a}

\item{file_name}{a}

\item{path}{a}

\item{translate}{a}

\item{digits}{a}

\item{percent}{a}

\item{accuracy}{a}

\item{override}{a}
}
\description{
Title
}
\examples{
a
}
