% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_lyx.R
\name{save_to_lyx}
\alias{save_to_lyx}
\title{Title}
\usage{
save_to_lyx(
  currentValue,
  currentName,
  latexFile,
  path = NULL,
  translate = TRUE,
  digits = 2,
  percent = FALSE,
  accuracy = 1,
  override = TRUE
)
}
\arguments{
\item{currentValue}{a}

\item{currentName}{a}

\item{latexFile}{a}

\item{path}{a}

\item{translate}{a}

\item{digits}{a}

\item{percent}{a}

\item{accuracy}{a}

\item{override}{a}
}
\description{
Title
}
\examples{
a
}
