% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savetexvalue.R
\docType{package}
\name{savetexvalue}
\alias{savetexvalue}
\title{savetexvalue: A package for creation and management of value-tex files}
\description{
The savetexvalue package provides functions for creation and management of
special formatted '.tex' files containing values calculated in R. The files
contain latex commands with user provided names and values that allow
integrating these values in a latex project in an automatic fashion instead
of copy-pasting.
}
\section{savetexvalue functions}{
 The savetexvalue package contains three
functions: \itemize{ \item \code{save_tex_value} for saving values to file.
\item \code{remove_tex_value} for removing values from file. \item
\code{print_tex_value} for displaying the contents of a value-tex file in
your R session. }
}

\author{
Ro'ee Levy, Ori Shoham \cr Maintainer: Ori Shoham
\email{ori.shoham1@gmail.com}
}
