% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tex_value.R
\name{format_values}
\alias{format_values}
\title{Formats values for lyx}
\usage{
format_values(values, percent, accuracy, digits, translate)
}
\arguments{
\item{values}{a}

\item{percent}{a}

\item{accuracy}{a}

\item{digits}{a}

\item{translate}{a}
}
\value{
A formatted character vector.
}
\description{
Formats values for lyx
}
\keyword{internal}
