% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tex_value.R
\name{handle_file_name}
\alias{handle_file_name}
\title{Title}
\usage{
handle_file_name(file_name, path = NULL)
}
\arguments{
\item{file_name}{a}

\item{path}{a}
}
\value{
a complete file name (with path and extension)
}
\description{
Title
}
\keyword{internal}
