% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tex_value.R
\name{handle_file_name}
\alias{handle_file_name}
\title{Constructs fully qualified file name}
\usage{
handle_file_name(file_name, path = NULL)
}
\arguments{
\item{file_name}{File name to create on disk. \code{file_name} should have
either a \code{.tex} extension or no extension at all.}

\item{path}{Path to the directory to save the file to. \code{path} and
\code{file_name} are combined to create fully qualified file name. If
\code{Null} and \code{file_name} does not specify a full file name defaults
to working directory.}
}
\value{
a complete file name (with path and extension)
}
\description{
\code{hanle_file_name} checks the extension of \code{file_name}, breaks if it
is not a '.tex' extension and adds '.tex' extension if no extension exists.
If \code{path} is not \code{NULL} combines \code{path} and \code{file_name}.
}
\keyword{internal}
