% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tex_value.R
\name{print_tex_value}
\alias{print_tex_value}
\title{Title}
\usage{
print_tex_value(file_name, path = NULL, names = NULL)
}
\arguments{
\item{file_name}{a}

\item{path}{a}

\item{names}{ab}
}
\value{

}
\description{
Title
}
\examples{
a
}
