% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tex_value.R
\name{remove_tex_value}
\alias{remove_tex_value}
\title{Remove certain values from a values-tex file}
\usage{
remove_tex_value(names, file_name, path = NULL)
}
\arguments{
\item{names}{Character vector of names specifying which values are to be
removed.}

\item{file_name}{File name containing values to be removed. file_name should
have either a .tex extension or no extension at all.}

\item{path}{Path to the directory to save the file to. \code{path} and
\code{file_name} are combined to create fully qualified file name. If
\code{Null} and \code{file_name} does not specify a full file name defaults
to working directory.}
}
\description{
Remove certain values from a values-tex file
}
\examples{
\dontrun{
file <- tempfile()
save_tex_value(values = 1:3, names = c("a","b","c"), file_name = file)

# remove "a" and "b" from file
remove_tex_value(file_name = paste0(file, ".tex"), names = c("a", "b"))

# delete file with base::unlink()
unlink(paste0(file,".tex"))

}
}
