% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tex_values.R
\name{remove_tex_value}
\alias{remove_tex_value}
\title{Title}
\usage{
remove_tex_value(names, file_name, path = NULL)
}
\arguments{
\item{names}{Character vector of names. Each name must be unique and can
consist of letters only.}

\item{file_name}{File name to create on disk. \code{file_name} should have
either a \code{.tex} extension or no extension at all.}

\item{path}{Path to the directory to save the file to. \code{path} and
\code{file_name} are combined to create fully qualified file name. If
\code{Null} and \code{file_name} does not specify a full file name defaults
to working directory.}
}
\description{
Title
}
\examples{
a

}
